Filter the data that are retrieved from the API using a Filter stateless component:
It has have event listeners adequately placed in it.
It uses lifting to pass user data to the parent components.
