+++ src/components/SearchForm.js
@@ -1,41 +1,22 @@
-import { useState } from 'react';
-import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { FaSearch } from 'react-icons/fa';
-import { filterCity } from '../redux/actions';
 
-export const SearchForm = ({ getSearchQuery }) => {
-  const [state, setState] = useState({ searchQuery: '' });
-  const handleSearchQueryChange = event => {
-    event.preventDefault();
-    setState({ searchQuery: event.target.value });
-    getSearchQuery(event.target.value);
-  };
-  const handleSearchQuerySubmit = event => {
-    event.preventDefault();
-    getSearchQuery(state.searchQuery);
-  };
-  return (
-    <form onSubmit={event => handleSearchQuerySubmit(event)}>
-      <input
-        type="text"
-        onChange={event => handleSearchQueryChange(event)}
-        value={state.searchQuery}
-        placeholder="City or State"
-      />
-      <button type="submit">
-        <FaSearch />
-      </button>
-    </form>
-  );
-};
+export const SearchForm = ({ handleSearchQueryChange, handleSearchQuerySubmit }) => (
+  <form onSubmit={event => handleSearchQuerySubmit(event)}>
+    <input
+      type="text"
+      onChange={event => handleSearchQueryChange(event)}
+      placeholder="City or State"
+    />
+    <button type="submit">
+      <FaSearch />
+    </button>
+  </form>
+);
 
 SearchForm.propTypes = {
-  getSearchQuery: PropTypes.func.isRequired,
+  handleSearchQueryChange: PropTypes.func.isRequired,
+  handleSearchQuerySubmit: PropTypes.func.isRequired,
 };
 
-const mapDispatchToProps = dispatch => ({
-  getSearchQuery: searchQuery => dispatch(filterCity(searchQuery)),
-});
-
-export default connect(null, mapDispatchToProps)(SearchForm);
+export default SearchForm;
