+++ src/containers/NavBar.js
@@ -2,13 +2,29 @@
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { FaHome } from 'react-icons/fa';
+import { useState } from 'react';
+import { connect } from 'react-redux';
 import SearchForm from '../components/SearchForm';
+import { filterCity } from '../redux/actions';
 
-const NavBar = ({ renderedComponent }) => {
+const NavBar = ({ renderedComponent, getSearchQuery }) => {
+  const [state, setState] = useState({ searchQuery: '' });
+  const handleSearchQueryChange = event => {
+    event.preventDefault();
+    setState({ searchQuery: event.target.value });
+    getSearchQuery(event.target.value);
+  };
+  const handleSearchQuerySubmit = event => {
+    event.preventDefault();
+    getSearchQuery(state.searchQuery);
+  };
   const isFormRendered = renderedComponent === 'all-cities' ? (
     <div className="nav">
       <div>
-        <SearchForm />
+        <SearchForm
+          handleSearchQueryChange={handleSearchQueryChange}
+          handleSearchQuerySubmit={handleSearchQuerySubmit}
+        />
       </div>
     </div>
   ) : '';
@@ -27,6 +43,11 @@ const NavBar = ({ renderedComponent }) => {
 
 NavBar.propTypes = {
   renderedComponent: PropTypes.string.isRequired,
+  getSearchQuery: PropTypes.func.isRequired,
 };
 
-export default NavBar;
+const mapDispatchToProps = dispatch => ({
+  getSearchQuery: searchQuery => dispatch(filterCity(searchQuery)),
+});
+
+export default connect(null, mapDispatchToProps)(NavBar);
